Refactor and expand ndarray test suite to improve coverage and robustness
This commit completes a comprehensive 5-phase test coverage improvement plan, elevating global package coverage from 60.67% Line / 57.80% Branch to 64.44% Line / 62.03% Branch, and verifying all FFI boundary conditions, type promotions, strided layouts, and safety preconditions.

Summary of Changes:
- Phase 1: Established a dedicated spacers_test.dart and exhaustively covered spacer grid constructors (linspace, logspace, geomspace and their grid equivalents) across all 8 FFI-bound DTypes.
- Phase 2: Expanded arithmetic and trigonometric tests. Added native C++ complex hyperbolic support (sinh, cosh, tanh, asinh) in custom_ufuncs.cpp and bound it to FFI. Fixed a type promotion bug in helpers for uint8 + int16.
- Phase 3: Expanded stats tests with NaN-aware reductions and strided medians/quantiles. Fixed a critical striding bug in nanmean where global reductions bypassed strides/offsets.
- Phase 4: Implemented the missing argwhere operator in C++ and Dart. Added FFI dispatches for complex count_nonzero.
- Phase 5: Expanded FFT and Calculus tests. Extended the FFT API to allow output buffers to drive target precision, resolving a static FFI unreachability gap. Added complex coordinate/step integrations in trapz.
- Stability: Fixed flakiness in Poisson, Normal, and Binomial random distribution tests by adding fixed seeds to all statistical validation tests, making them 100% deterministic.

TAG=agy
CONV=523dfab0-8eba-4aae-b392-51f2da05eac8

+++ pkgs/ndarray/test/regression/regression_test.dart
@@ -170,8 +170,8 @@ void main() {
       expect(c.toList(), [2, 2, 3]);
       expect(
         c.dtype,
-        DType.int32,
-      ); // resolved dtype of uint8 and int16 is int32
+        DType.int16,
+      ); // resolved dtype of uint8 and int16 is int16 (NumPy style)
     });
 
     test('floor_divide with uint8/float64', () {
@@ -191,7 +191,7 @@ void main() {
       // This should not crash
       final c = remainder(a, b);
       expect(c.toList(), [2, 0, 1]);
-      expect(c.dtype, DType.int32);
+      expect(c.dtype, DType.int16);
     });
 
     test('remainder with uint8/float64', () {
